pyproject.toml
[tool.poetry]
name = "microlibraryservices"
version = "0.1.0"
description = "Dependency mgmt for microLibraryServices"
authors = ["tandukar <84461389+tandukar@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
amqp = "5.2.0"
asgiref = "3.7.2"
async-timeout = "4.0.3"
billiard = "4.2.0"
celery = "5.3.6"
click = "8.1.7"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorama = "0.4.6"
django = "4.2.11"
djangorestframework = "3.15.0"
djangorestframework-simplejwt = "5.3.1"
kombu = "5.3.5"
pika = "1.3.2"
prompt-toolkit = "3.0.43"
pyjwt = "2.8.0"
python-dateutil = "2.9.0.post0"
python-decouple = "3.8"
redis = "5.0.3"
six = "1.16.0"
sqlparse = "0.4.4"
typing-extensions = "4.10.0"
tzdata = "2024.1"
vine = "5.1.0"
wcwidth = "0.2.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
